+++ autograder-core/src/main/java/de/firemage/autograder/core/integrated/effects/AssignmentEffect.java
@@ -1,0 +1,7 @@
+package de.firemage.autograder.core.integrated.effects;
+
+import spoon.reflect.reference.CtVariableReference;
+
+public interface AssignmentEffect extends Effect {
+    CtVariableReference<?> target();
+}
