+++ autograder-core/src/main/java/de/firemage/autograder/core/integrated/effects/TerminalEffect.java
@@ -1,0 +1,4 @@
+package de.firemage.autograder.core.integrated.effects;
+
+public interface TerminalEffect extends Effect {
+}
